Add support for virtual-machines, group by tags
Virtual machines code based on https://github.com/AAbouZaid/netbox-as-ansible-inventory/pull/15. Added a little tweak to make it group by device roles correctly.

Group by tags code based on https://github.com/AAbouZaid/netbox-as-ansible-inventory/issues/35. Adjusted it to the most recent version of the script and added and option for this in the yaml file.